requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
arrow==1.3.0
asgiref==3.8.1
binaryornot==0.4.4
blinker==1.9.0
CacheControl==0.14.1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cookiecutter==2.6.0
cryptography==43.0.3
Django==5.1.3
fastapi==0.115.5
firebase-admin==6.6.0
Flask==3.1.0
flet==0.22.1
flet-core==0.22.1
flet-runtime==0.22.1
gcloud==0.18.3
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.68.0
grpcio-status==1.68.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jwcrypto==1.5.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msgpack==1.1.0
numpy==2.1.3
oauth2client==4.1.3
oauthlib==3.2.2
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==23.2
pillow==11.0.0
proto-plus==1.25.0
protobuf==5.29.0rc3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.1
pydantic_core==2.27.1
Pygments==2.18.0
PyJWT==2.10.0
pyparsing==3.2.0
pypng==0.20220715.0
Pyrebase4==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jwt==4.1.0
python-slugify==8.0.4
PyYAML==6.0.2
qrcode==7.4.2
repath==0.9.0
requests==2.32.3
requests-toolbelt==0.10.1
rich==13.9.4
rsa==4.9
setuptools==75.6.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.2
starlette==0.41.3
text-unidecode==1.3
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==1.26.20
uvicorn==0.32.1
watchdog==4.0.2
watchfiles==0.24.0
websockets==14.1
Werkzeug==3.1.3
